+++ tests/broker.test.js
@@ -4,6 +4,12 @@ describe('Broker class testing', () => {
 
   describe('Test Broker construction', () => {
 
+    test('instance of broker will be a child of the EventEmitter class', () => {
+      const broker = new Broker();
+      const {EventEmitter} = require('events');
+      expect(broker instanceof EventEmitter);
+    });
+
     test('instance of Broker class will have exchange and valid properties', () => {
       const broker = new Broker();
       expect(broker).toHaveProperty('exchange');
@@ -17,7 +23,18 @@ describe('Broker class testing', () => {
 
     test('instance of Broker class will be valid if exchange instance is passed to its constructor', () => {
       const Exchange = require('../src/exchange');
-      const broker = new Broker()
+      const credentials = {key: 'myKey', secret: 'mySecret', passphrase: 'myPassphrase'};
+      const exchange = new Exchange(credentials);
+      const broker = new Broker(exchange);
+      expect(broker.valid).toBe(true);
+    });
+
+    test('instances of Broker class are not valid if an invalid exchange instance is passed to the constructor', () => {
+      const Exchange = require('../src/exchange');
+      const credentials = {key: 'myKey'};
+      const exchange = new Exchange(credentials);
+      const broker = new Broker(exchange);
+      expect(broker.valid).toBe(false);
     });
   });
 });